Namespace all rename code inside crazyfun mappings
